+++ modules/juce_gui_basics/native/juce_mac_CGMetalLayerRenderer.h
@@ -341,7 +341,7 @@ private:
 
                 if (alignedSize > size)
                 {
-                    size = std::max (alignedSize, alignTo ((size_t) (size * growthFactor), pagesize));
+                    size = std::max (alignedSize, alignTo ((size_t) (static_cast<float>(size) * growthFactor), pagesize));
                     allocation = std::make_unique<AllocationWrapper> (pagesize, size);
                 }
